+++ packages/pwa/app/utils/routes-utils.test.js
@@ -27,45 +27,50 @@ describe('configureRoutes', function() {
     afterEach(() => {
         process.env = env
     })
+    const CompA = () => <div>This is component A</div>
+    const CompC = () => <div>This is component C</div>
+
+    const routes = [
+        {
+            path: '/',
+            component: CompA,
+            exact: true
+        },
+        {
+            path: '/category/:categoryId',
+            component: CompC,
+            exact: true
+        }
+    ]
     test('should return all permutation of path including site and locales ', () => {
-        getConfig.mockImplementation(() => mockConfig.app.hosts)
-        const CompA = () => <div>This is component A</div>
-        const CompC = () => <div>This is component C</div>
-
-        // mock current domain/host
-        process.env.EXTERNAL_DOMAIN_NAME = 'www.domain-1.com'
-
-        const routes = [
-            {
-                path: '/',
-                component: CompA,
-                exact: true
-            },
-            {
-                path: '/category/:categoryId',
-                component: CompC,
-                exact: true
-            }
-        ]
-        const configuredRoutes = configureRoutes(routes, {ignoredRoutes: '/'})
+        getConfig.mockImplementation(() => mockConfig)
+        const configuredRoutes = configureRoutes(routes, {ignoredRoutes: ['/']})
         expect(configuredRoutes[configuredRoutes.length - 1].path).toEqual('/category/:categoryId')
-        expect(configuredRoutes.length).toEqual(61)
+        expect(configuredRoutes.length).toEqual(81)
         const paths = configuredRoutes.map((route) => route.path)
         expect(paths).toEqual(expectedPathsResult)
+    })
+
+    test('should return the origin routes', () => {
+        getConfig.mockImplementation(() => mockConfig)
+        const configuredRoutes = configureRoutes(routes, {
+            ignoredRoutes: ['/', '/category/:categoryId']
+        })
+        expect(configuredRoutes.length).toEqual(2)
     })
 })
 
 const expectedPathsResult = [
     '/',
-    '/RefArch/en-US/category/:categoryId',
-    '/RefArch/en/category/:categoryId',
-    '/RefArch/fr-FR/category/:categoryId',
-    '/RefArch/fr/category/:categoryId',
-    '/RefArch/it-IT/category/:categoryId',
-    '/RefArch/it/category/:categoryId',
-    '/RefArch/en-HK/category/:categoryId',
-    '/RefArch/en-GB/category/:categoryId',
-    '/RefArch/uk/category/:categoryId',
+    '/site-1/en-US/category/:categoryId',
+    '/site-1/en/category/:categoryId',
+    '/site-1/fr-FR/category/:categoryId',
+    '/site-1/fr/category/:categoryId',
+    '/site-1/it-IT/category/:categoryId',
+    '/site-1/it/category/:categoryId',
+    '/site-1/en-HK/category/:categoryId',
+    '/site-1/en-GB/category/:categoryId',
+    '/site-1/uk/category/:categoryId',
     '/us/en-US/category/:categoryId',
     '/us/en/category/:categoryId',
     '/us/fr-FR/category/:categoryId',
@@ -75,6 +80,15 @@ const expectedPathsResult = [
     '/us/en-HK/category/:categoryId',
     '/us/en-GB/category/:categoryId',
     '/us/uk/category/:categoryId',
+    '/site-2/en-US/category/:categoryId',
+    '/site-2/en/category/:categoryId',
+    '/site-2/fr-FR/category/:categoryId',
+    '/site-2/fr/category/:categoryId',
+    '/site-2/it-IT/category/:categoryId',
+    '/site-2/it/category/:categoryId',
+    '/site-2/en-HK/category/:categoryId',
+    '/site-2/en-GB/category/:categoryId',
+    '/site-2/uk/category/:categoryId',
     '/eu/en-US/category/:categoryId',
     '/eu/en/category/:categoryId',
     '/eu/fr-FR/category/:categoryId',
@@ -84,15 +98,24 @@ const expectedPathsResult = [
     '/eu/en-HK/category/:categoryId',
     '/eu/en-GB/category/:categoryId',
     '/eu/uk/category/:categoryId',
-    '/RefArchGlobal/en-US/category/:categoryId',
-    '/RefArchGlobal/en/category/:categoryId',
-    '/RefArchGlobal/fr-FR/category/:categoryId',
-    '/RefArchGlobal/fr/category/:categoryId',
-    '/RefArchGlobal/it-IT/category/:categoryId',
-    '/RefArchGlobal/it/category/:categoryId',
-    '/RefArchGlobal/en-HK/category/:categoryId',
-    '/RefArchGlobal/en-GB/category/:categoryId',
-    '/RefArchGlobal/uk/category/:categoryId',
+    '/site-3/en-US/category/:categoryId',
+    '/site-3/en/category/:categoryId',
+    '/site-3/fr-FR/category/:categoryId',
+    '/site-3/fr/category/:categoryId',
+    '/site-3/it-IT/category/:categoryId',
+    '/site-3/it/category/:categoryId',
+    '/site-3/en-HK/category/:categoryId',
+    '/site-3/en-GB/category/:categoryId',
+    '/site-3/uk/category/:categoryId',
+    '/site-4/en-US/category/:categoryId',
+    '/site-4/en/category/:categoryId',
+    '/site-4/fr-FR/category/:categoryId',
+    '/site-4/fr/category/:categoryId',
+    '/site-4/it-IT/category/:categoryId',
+    '/site-4/it/category/:categoryId',
+    '/site-4/en-HK/category/:categoryId',
+    '/site-4/en-GB/category/:categoryId',
+    '/site-4/uk/category/:categoryId',
     '/global/en-US/category/:categoryId',
     '/global/en/category/:categoryId',
     '/global/fr-FR/category/:categoryId',
@@ -102,10 +125,12 @@ const expectedPathsResult = [
     '/global/en-HK/category/:categoryId',
     '/global/en-GB/category/:categoryId',
     '/global/uk/category/:categoryId',
-    '/RefArch/category/:categoryId',
+    '/site-1/category/:categoryId',
     '/us/category/:categoryId',
+    '/site-2/category/:categoryId',
     '/eu/category/:categoryId',
-    '/RefArchGlobal/category/:categoryId',
+    '/site-3/category/:categoryId',
+    '/site-4/category/:categoryId',
     '/global/category/:categoryId',
     '/en-US/category/:categoryId',
     '/en/category/:categoryId',
@@ -119,115 +144,86 @@ const expectedPathsResult = [
     '/category/:categoryId'
 ]
 const mockConfig = {
-    app: {
-        defaultSite: 'RefArch',
-        url: {
-            site: 'none',
-            locale: 'none'
-        },
-        hosts: [
-            {
-                domain: 'www.domain-1.com',
-                defaultSite: 'RefArch',
-                sites: [
+    defaultSite: 'RefArch',
+    url: {
+        site: 'none',
+        locale: 'none'
+    },
+    sites: [
+        {
+            id: 'site-1',
+            alias: 'us',
+            l10n: {
+                supportedCurrencies: ['USD'],
+                defaultCurrency: 'USD',
+                defaultLocale: 'en-US',
+                supportedLocales: [
                     {
-                        id: 'RefArch',
-                        alias: 'us',
-                        l10n: {
-                            supportedCurrencies: ['USD'],
-                            defaultCurrency: 'USD',
-                            defaultLocale: 'en-US',
-                            supportedLocales: [
-                                {
-                                    id: 'en-US',
-                                    alias: 'en',
-                                    preferredCurrency: 'USD'
-                                }
-                            ]
-                        }
-                    },
-                    {
-                        id: 'RefArch',
-                        alias: 'eu',
-                        l10n: {
-                            supportedCurrencies: ['USD'],
-                            defaultCurrency: 'USD',
-                            supportedLocales: [
-                                {
-                                    id: 'fr-FR',
-                                    alias: 'fr',
-                                    preferredCurrency: 'USD'
-                                },
-                                {
-                                    id: 'it-IT',
-                                    alias: 'it',
-                                    preferredCurrency: 'USD'
-                                }
-                            ]
-                        }
-                    },
-                    {
-                        id: 'RefArch',
-                        defaultLocale: 'en-US',
-                        l10n: {
-                            supportedCurrencies: ['USD'],
-                            defaultCurrency: 'USD',
-                            supportedLocales: [
-                                {
-                                    id: 'en-US',
-                                    preferredCurrency: 'USD'
-                                },
-                                {
-                                    id: 'en-HK',
-                                    preferredCurrency: 'HKD'
-                                }
-                            ],
-                            defaultLocale: 'en-US'
-                        }
-                    },
-                    {
-                        id: 'RefArchGlobal',
-                        alias: 'global',
-                        l10n: {
-                            supportedCurrencies: ['GBP', 'EUR'],
-                            defaultCurrency: 'GBP',
-                            supportedLocales: [
-                                {
-                                    id: 'en-GB',
-                                    alias: 'uk',
-                                    preferredCurrency: 'GBP'
-                                },
-                                {
-                                    id: 'fr-FR',
-                                    alias: 'fr',
-                                    preferredCurrency: 'EUR'
-                                }
-                            ]
-                        }
-                    }
-                ]
-            },
-            {
-                domain: 'www.domain-2.com',
-                defaultSite: 'site-de',
-                sites: [
-                    {
-                        id: 'site-de',
-                        l10n: {
-                            supportedCurrencies: ['EUR'],
-                            defaultCurrency: 'EUR',
-                            defaultLocale: 'de-DE',
-                            supportedLocales: [
-                                {
-                                    id: 'de-DE',
-                                    alias: 'de',
-                                    preferredCurrency: 'EUR'
-                                }
-                            ]
-                        }
+                        id: 'en-US',
+                        alias: 'en',
+                        preferredCurrency: 'USD'
                     }
                 ]
             }
-        ]
-    }
+        },
+        {
+            id: 'site-2',
+            alias: 'eu',
+            l10n: {
+                supportedCurrencies: ['USD'],
+                defaultCurrency: 'USD',
+                supportedLocales: [
+                    {
+                        id: 'fr-FR',
+                        alias: 'fr',
+                        preferredCurrency: 'USD'
+                    },
+                    {
+                        id: 'it-IT',
+                        alias: 'it',
+                        preferredCurrency: 'USD'
+                    }
+                ]
+            }
+        },
+        {
+            id: 'site-3',
+            defaultLocale: 'en-US',
+            l10n: {
+                supportedCurrencies: ['USD'],
+                defaultCurrency: 'USD',
+                supportedLocales: [
+                    {
+                        id: 'en-US',
+                        preferredCurrency: 'USD'
+                    },
+                    {
+                        id: 'en-HK',
+                        preferredCurrency: 'HKD'
+                    }
+                ],
+                defaultLocale: 'en-US'
+            }
+        },
+        {
+            id: 'site-4',
+            alias: 'global',
+            l10n: {
+                supportedCurrencies: ['GBP', 'EUR'],
+                defaultCurrency: 'GBP',
+                supportedLocales: [
+                    {
+                        id: 'en-GB',
+                        alias: 'uk',
+                        preferredCurrency: 'GBP'
+                    },
+                    {
+                        id: 'fr-FR',
+                        alias: 'fr',
+                        preferredCurrency: 'EUR'
+                    }
+                ]
+            }
+        }
+    ]
 }